+++ assets/js/app/core/layout/01_layout-controllers.js
@@ -129,7 +129,7 @@
 
               $scope.alerts = [];
 
-              if((UserService.user().node.id == node.id ) || node.checkingConnection) return false;
+              if((UserService.user().node && UserService.user().node.id == node.id ) || node.checkingConnection) return false;
 
 
               // Check if the connection is valid